Admin CSS
This makes the options panel fit on the screen on smaller devices and removes the sidebar to conform with the style of other wordpress options in mobile view.

The CSS file is just the edited SCSS file converted. This is my first contribution so I wasn't sure on the practice here.